split strings.cs into separate classes
