+++ DragonFruit.Six.API/Data/Deserializers/OperatorStatsDeserializer.cs
@@ -19,20 +19,20 @@ namespace DragonFruit.Six.API.Data.Deserializers
             {
                 op.Guid = guid;
 
-                op.Kills = json.GetUInt(string.Format(OverallOperator.Kills, op.Index));
-                op.Deaths = json.GetUInt(string.Format(OverallOperator.Deaths, op.Index));
-                op.Kd = json.GetFloat(string.Format(OverallOperator.Kills, op.Index), 1) /
-                        json.GetFloat(string.Format(OverallOperator.Deaths, op.Index), 1);
+                op.Kills = json.GetUInt(string.Format(Operator.Kills, op.Index));
+                op.Deaths = json.GetUInt(string.Format(Operator.Deaths, op.Index));
+                op.Kd = json.GetFloat(string.Format(Operator.Kills, op.Index), 1) /
+                        json.GetFloat(string.Format(Operator.Deaths, op.Index), 1);
 
-                op.Wins = json.GetUInt(string.Format(OverallOperator.Wins, op.Index));
-                op.Losses = json.GetUInt(string.Format(OverallOperator.Losses, op.Index));
-                op.Wl = json.GetFloat(string.Format(OverallOperator.Wins, op.Index), 1) /
-                        json.GetFloat(string.Format(OverallOperator.Losses, op.Index), 1);
+                op.Wins = json.GetUInt(string.Format(Operator.Wins, op.Index));
+                op.Losses = json.GetUInt(string.Format(Operator.Losses, op.Index));
+                op.Wl = json.GetFloat(string.Format(Operator.Wins, op.Index), 1) /
+                        json.GetFloat(string.Format(Operator.Losses, op.Index), 1);
 
-                op.RoundsPlayed = json.GetUInt(string.Format(OverallOperator.Rounds, op.Index));
-                op.TimePlayed = TimeSpan.FromSeconds(json.GetUInt(string.Format(OverallOperator.Time, op.Index)));
-                op.Headshots = json.GetUInt(string.Format(OverallOperator.Headshots, op.Index));
-                op.Downs = json.GetUInt(string.Format(OverallOperator.Downs, op.Index));
+                op.RoundsPlayed = json.GetUInt(string.Format(Operator.Rounds, op.Index));
+                op.TimePlayed = TimeSpan.FromSeconds(json.GetUInt(string.Format(Operator.Time, op.Index)));
+                op.Headshots = json.GetUInt(string.Format(Operator.Headshots, op.Index));
+                op.Downs = json.GetUInt(string.Format(Operator.Downs, op.Index));
 
                 op.ActionCount = json.GetUInt(op.OperatorActionResultId);
