+++ DragonFruit.Six.API/Data/Strings/Accounts.cs
@@ -1,0 +1,15 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class Accounts
+    {
+        public static readonly string Name = "nameOnPlatform";
+        public static readonly string Platform = "platformType";
+
+        public static readonly string PlatformIdentifier = "idOnPlatform";
+        public static readonly string ProfileIdentifier = "profileId";
+        public static readonly string UserIdentifier = "userId";
+    }
+}
