+++ DragonFruit.Six.API/Data/Strings/General.cs
@@ -1,0 +1,26 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class General
+    {
+        public static readonly string Barricades = "generalpvp_barricadedeployed:infinite";
+        public static readonly string Reinforcements = "generalpvp_reinforcementdeploy:infinite";
+
+        public static readonly string Downs = "generalpvp_dbno:infinite";
+        public static readonly string Revives = "generalpvp_revive:infinite";
+
+        public static readonly string Headshots = "generalpvp_headshot:infinite";
+        public static readonly string Knives = "generalpvp_meleekills:infinite";
+        public static readonly string Penetrations = "generalpvp_penetrationkills:infinite";
+
+        public static readonly string Assists = "generalpvp_killassists:infinite";
+        public static readonly string Suicides = "generalpvp_suicide:infinite";
+
+        public static readonly string BulletFired = "generalpvp_bulletfired:infinite";
+        public static readonly string BulletHit = "generalpvp_bullethit:infinite";
+
+        public static readonly string Level = "level";
+    }
+}
