+++ DragonFruit.Six.API/Data/Strings/GeneralCasual.cs
@@ -1,0 +1,14 @@
+﻿namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class GeneralCasual
+    {
+        public static readonly string Kills = "casualpvp_kills:infinite";
+        public static readonly string Deaths = "casualpvp_death:infinite";
+
+        public static readonly string Wins = "casualpvp_matchwon:infinite";
+        public static readonly string Losses = "casualpvp_matchlost:infinite";
+
+        public static readonly string MatchesPlayed = "casualpvp_matchplayed:infinite";
+        public static readonly string Time = "casualpvp_timeplayed:infinite";
+    }
+}
