+++ DragonFruit.Six.API/Data/Strings/GeneralRanked.cs
@@ -1,0 +1,17 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class GeneralRanked
+    {
+        public static readonly string Kills = "rankedpvp_kills:infinite";
+        public static readonly string Deaths = "rankedpvp_death:infinite";
+
+        public static readonly string Wins = "rankedpvp_matchwon:infinite";
+        public static readonly string Losses = "rankedpvp_matchlost:infinite";
+
+        public static readonly string MatchesPlayed = "rankedpvp_matchplayed:infinite";
+        public static readonly string Time = "rankedpvp_timeplayed:infinite";
+    }
+}
