+++ DragonFruit.Six.API/Data/Strings/GeneralTraining.cs
@@ -1,0 +1,16 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class GeneralTraining
+    {
+        public static readonly string Kills = "generalpve_kills:infinite";
+        public static readonly string Deaths = "generalpve_death:infinite";
+
+        public static readonly string Wins = "generalpve_matchwon:infinite";
+        public static readonly string Losses = "generalpve_matchlost:infinite";
+
+        public static readonly string Time = "generalpve_timeplayed:infinite";
+    }
+}
