+++ DragonFruit.Six.API/Data/Strings/Highscores.cs
@@ -1,0 +1,12 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class Highscores
+    {
+        public static readonly string Bomb = "plantbombpvp_bestscore:infinite";
+        public static readonly string Hostage = "rescuehostagepvp_bestscore:infinite";
+        public static readonly string Secure = "secureareapvp_bestscore:infinite";
+    }
+}
