+++ DragonFruit.Six.API/Data/Strings/Login.cs
@@ -1,0 +1,14 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class Login
+    {
+        public static readonly string Sessions = "sessionsCount";
+        public static readonly string Guid = "profileId";
+        public static readonly string PlatformId = "applicationId";
+        public static readonly string FirstLogin = "firstSessionDate";
+        public static readonly string LastLogin = "lastSessionDate";
+    }
+}
