+++ DragonFruit.Six.API/Data/Strings/Misc.cs
@@ -1,0 +1,12 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class Misc
+    {
+        public static readonly string Results = "results";
+        public static readonly string Profile = "profiles";
+        public static readonly string Players = "players";
+    }
+}
