+++ DragonFruit.Six.API/Data/Strings/Operator.cs
@@ -1,0 +1,17 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class Operator
+    {
+        public static readonly string Wins = "operatorpvp_roundwon:{0}:infinite";
+        public static readonly string Losses = "operatorpvp_roundlost:{0}:infinite";
+        public static readonly string Kills = "operatorpvp_kills:{0}:infinite";
+        public static readonly string Deaths = "operatorpvp_death:{0}:infinite";
+        public static readonly string Downs = "operatorpvp_headshot:{0}:infinite";
+        public static readonly string Headshots = "operatorpvp_dbno:{0}:infinite";
+        public static readonly string Rounds = "operatorpvp_roundplayed:{0}:infinite";
+        public static readonly string Time = "operatorpvp_timeplayed:{0}:infinite";
+    }
+}
