+++ DragonFruit.Six.API/Data/Strings/OverallMultiplayer.cs
@@ -1,0 +1,16 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class OverallMultiplayer
+    {
+        public static readonly string Kills = "generalpvp_kills:infinite";
+        public static readonly string Deaths = "generalpvp_death:infinite";
+
+        public static readonly string Wins = "generalpvp_matchwon:infinite";
+        public static readonly string Losses = "generalpvp_matchlost:infinite";
+
+        public static readonly string Time = "generalpvp_timeplayed:infinite";
+    }
+}
