+++ DragonFruit.Six.API/Data/Strings/Weapon.cs
@@ -1,0 +1,13 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Data.Strings
+{
+    public static class Weapon
+    {
+        public static readonly string Kills = "weapontypepvp_kills:{0}:infinite";
+        public static readonly string Headshots = "weapontypepvp_headshot:{0}:infinite";
+        public static readonly string ShotsFired = "weapontypepvp_bulletfired:{0}:infinite";
+        public static readonly string ShotsHit = "weapontypepvp_bullethit:{0}:infinite";
+    }
+}
